cleaned up `add_subscription`
If you started to add a new podcast subscription, and changed your mind, there was no way to cleanly exit out of it.  It would add a new subscription and then you'd have to go back and delete it.  I just added a way to exit out of the function if you realized you made a mistake.